Flutter Card Details Widget
This Flutter widget is designed to display detailed information about a card. It is used to present various attributes and features of the card, including its type, name, level, description, stats (attack and defense), race, attribute, archetype, formats, and pricing information. The widget provides a visually appealing and organized layout with customizable styles and formatting. It also includes buttons for viewing the card in different sets and displays the current prices of the card from various sources such as Cardmarket, TCGPlayer, eBay, Amazon, and Coolstuffinc. The widget is flexible and can be easily integrated into Flutter applications for card-related functionalities and displays.