Customizable Card Widgets
The "CardNameAndLevel" widget enables the creation of customizable cards that display a name and an optional level. With parameters for name, level, and level asset, this widget generates cards with dynamic text and an optional level indicator using the "CardLevel" widget.

The "CardRaceAndAttribute" widget is another customizable card widget that showcases the race and attribute of a character. By providing the race and attribute parameters, this widget generates visually appealing cards with race images fetched from assets. It also includes an optional attribute image using the "CardDetailWidget" component.