Remote Networking
This code snippet demonstrates the implementation of remote networking using the Dio package. The code defines an abstract class called RemoteClient with a single method get for making HTTP GET requests. The DioClient class implements the RemoteClient interface and uses the Dio package to send HTTP requests. It handles successful responses by returning the response data and throws a ServerException in case of Dio errors.