Type Adapter for CardMiscInfo.
The CardMiscInfoAdapter is a type adapter used for serializing and deserializing instances of the CardMiscInfo widget. It implements the TypeAdapter interface and provides methods to read and write binary data. The adapter is generated code and should not be modified manually. It maps the fields of CardMiscInfo to their corresponding positions in the binary data and handles the conversion between the binary format and the CardMiscInfo object. This adapter is used in conjunction with Hive.