CardOwned widget.
The CardOwned widget is a model class that represents an owned card. It is used to store information about a specific card, including the quantity owned, set code, edition (1st edition or unlimited), set name, and ID. This class implements the Equatable mixin to enable easy comparison of objects based on their properties. The class is annotated with @HiveType to enable serialization and deserialization using Hive, a NoSQL database for Flutter. The class also defines a key getter that combines the set code and edition to create a unique identifier for the card.